ignore non conf files in /etc/portage/repos.conf/
portage ignores any file in /etc/portage/repos.conf/ (when that is in
fact a directory) that does not have a name of the form "*.conf".

gentoo-bashcomp should emulate the behavior of portage. Otherwise you
might run into issues: e.g. if there is a backup of eselect-repo.conf
called eselect-repo.conf~ that contains references to a now removed
overlay, gentoo-bashcomp should not try to search for completions in the
now nonexistent repo directory