portageq envvar: preserve PATH to avoid command validation warnings
Some of Portage's environment variables undergo command validation.
When one is set to a relative value, Portage relies on PATH to find the
binary.
We need to preserve it after purging the environment (`env -i`) if we
want to avoid "<key> is invalid: <value>" warnings.

For Portage 3.0.63, this affects
PORTAGE_{B{,UN}ZIP2_COMMAND,LOG_FILTER_FILE_CMD}.

Bug: https://bugs.gentoo.org/588642
Signed-off-by: Lucio Sauer <watermanpaint@posteo.net>
Signed-off-by: Sam James <sam@gentoo.org>